wip quick_reject is quick!
This speeds up my demo slow simp proof by 4x, from about 1600ms to
about 400ms.

And this isn't even efficiently implemented!  It's traversing the
expr again and again, as well as each lemma again and again for
different exprs, to find these arg-heads.  The first of those, at
least, will be straightforward to fix.

But evidently the full `match`, which is what we'd try next, is
so slow that this is already a big win.